Streamlining characterization. (#25, #46)
